検索機能を最適化
